// src/router/types/routeNames.ts
export const enum RouteNames {
  // shared
  VACANCIES = 'VACANCIES',
  PROJECTS = 'PROJECTS',
  PROJECT = 'PROJECT',
  FAQ = 'FAQ',
  CONTACTS = 'CONTACTS',
  NOT_FOUND = 'NOT_FOUND',

  // vacancy
  VACANCY_DETAILS = 'VACANCY_DETAILS',

  // project
  PROJECT_DETAILS = 'PROJECT_DETAILS',
  PROJECT_VACANCIES = 'PROJECT_VACANCIES',

  // user
  USER_PAGE = 'USER_PAGE',
  USER_INFO = 'USER_INFO',
  // candidate
  RESPONSES = 'RESPONSES',
  // supervisor
  SUPERVISOR_PROJECT_VACANCIES = 'SUPERVISOR_PROJECT_VACANCIES',
  SUPERVISOR_VACANCY_RESPONSES = 'SUPERVISOR_VACANCY_RESPONSES',
  SUPERVISOR_VACANCY_RESPONSES_NEW = 'SUPERVISOR_VACANCY_RESPONSES_NEW',
  SUPERVISOR_VACANCY_RESPONSES_APPROVED = 'SUPERVISOR_VACANCY_RESPONSES_APPROVED',
  SUPERVISOR_VACANCY_RESPONSES_REJECTED = 'SUPERVISOR_VACANCY_RESPONSES_REJECTED',
  SUPERVISOR_PROJECTS = 'SUPERVISOR_PROJECTS',
  SUPERVISOR_VACANCY_CREATE = 'SUPERVISOR_VACANCY_CREATE',
  // admin
  ADMIN_VACANCY_OFFER = 'ADMIN_VACANCY_OFFER',
  ADMIN_SKILL_OFFER = 'ADMIN_SKILL_OFFER',
  ADMIN_PROJECT_CREATE = 'ADMIN_PROJECT_CREATE',
  PROJECTS_INTEREST = 'PROJECTS_INTEREST',
}